requirements.txt
aiofiles==0.6.0
aiogram==2.12.1
aiohttp==3.7.4.post0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
greenlet==1.0.0
h11==0.12.0
httptools==0.1.1
idna==3.1
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
psycopg2-binary==2.8.6
pydantic==1.8.1
PyJWT==2.0.1
python-dotenv==0.15.0
python-telegram-bot==13.4.1
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.4.2
starlette==0.13.6
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
yarl==1.6.3
